+++ crates/vent-common/src/window.rs
@@ -27,6 +27,6 @@ impl VentWindow {
             (rgba, width, height)
         };
         winit::window::Icon::from_rgba(icon_rgba, icon_width, icon_height)
-            .expect("Failed to open icon")
+            .expect("Failed to load icon")
     }
 }
