+++ crates/vent-runtime/src/render/camera/mod.rs
@@ -7,36 +7,11 @@ pub trait Camera {
     fn new() -> Self
     where
         Self: Sized;
-    // ugly i know :c
-    #[must_use]
-    fn build_view_matrix_2d(&mut self, aspect_ratio: f32) -> UBO2D;
-
-    #[must_use]
-    fn build_view_matrix_3d(&mut self, aspect_ratio: f32) -> UBO3D;
-}
-
-pub struct BasicCamera {
-    fovy: f32,
-    znear: f32,
-    zfar: f32,
-    pub rotation: glam::Quat,
-}
-
-impl Default for BasicCamera {
-    #[inline]
-    fn default() -> Self {
-        Self {
-            fovy: 60.0,
-            znear: 0.1,
-            zfar: 10000.0,
-            rotation: glam::Quat::IDENTITY,
-        }
-    }
 }
 
 pub struct Camera2D {
-    pub basic_cam: BasicCamera,
     pub position: glam::Vec2,
+    pub rotation: glam::Quat,
 }
 
 impl Camera for Camera2D {
@@ -47,25 +22,25 @@ impl Camera for Camera2D {
         Self: Sized,
     {
         Self {
-            basic_cam: BasicCamera::default(),
             position: glam::Vec2::ZERO,
+            rotation: glam::Quat::IDENTITY,
         }
     }
+}
 
+impl Camera2D {
     #[must_use]
-    fn build_view_matrix_2d(&mut self, _aspect_ratio: f32) -> UBO2D {
-        todo!()
-    }
-
-    #[must_use]
-    fn build_view_matrix_3d(&mut self, _aspect_ratio: f32) -> UBO3D {
+    pub fn build_view_matrix_2d(&mut self, _aspect_ratio: f32) -> UBO2D {
         todo!()
     }
 }
 
 pub struct Camera3D {
-    pub basic_cam: BasicCamera,
+    fovy: f32,
+    znear: f32,
+    zfar: f32,
     pub position: glam::Vec3,
+    pub rotation: glam::Quat,
 }
 
 impl Camera for Camera3D {
@@ -76,24 +51,20 @@ impl Camera for Camera3D {
         Self: Sized,
     {
         Self {
-            basic_cam: BasicCamera::default(),
+            fovy: 60.0,
+            znear: 0.1,
+            zfar: 10000.0,
+            rotation: glam::Quat::IDENTITY,
             position: Vec3::ZERO,
         }
     }
+}
 
+impl Camera3D {
     #[must_use]
-    fn build_view_matrix_2d(&mut self, _aspect_ratio: f32) -> UBO2D {
-        todo!()
-    }
-
-    #[must_use]
-    fn build_view_matrix_3d(&mut self, aspect_ratio: f32) -> UBO3D {
-        let projection = glam::Mat4::perspective_lh(
-            self.basic_cam.fovy.to_radians(),
-            aspect_ratio,
-            self.basic_cam.znear,
-            self.basic_cam.zfar,
-        );
+    pub fn build_view_matrix_3d(&mut self, aspect_ratio: f32) -> UBO3D {
+        let projection =
+            glam::Mat4::perspective_lh(self.fovy.to_radians(), aspect_ratio, self.znear, self.zfar);
 
         let view =
             glam::Mat4::look_at_lh(self.position, self.direction_from_rotation(), glam::Vec3::Y);
@@ -103,15 +74,13 @@ impl Camera for Camera3D {
             transformation: Mat4::IDENTITY.to_cols_array_2d(),
         }
     }
-}
 
-impl Camera3D {
     #[inline]
     #[must_use]
     fn direction_from_rotation(&self) -> glam::Vec3 {
-        let rot = self.basic_cam.rotation;
-        let cos_y = self.basic_cam.rotation.y.cos();
+        let (sin_pitch, cos_pitch) = self.rotation.y.sin_cos();
+        let (sin_yaw, cos_yaw) = self.rotation.x.sin_cos();
 
-        glam::vec3(rot.x.sin() * cos_y, rot.y.sin(), rot.x.cos() * cos_y)
+        glam::vec3(cos_pitch * cos_yaw, sin_pitch, cos_pitch * sin_yaw).normalize()
     }
 }
